funcionalidades ok

+++ banco/src/main/java/com/example/banco/services/ClienteServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.banco.services;
 import com.example.banco.DTO.ClienteDTO;
 import com.example.banco.entities.Cliente;
+import com.example.banco.exceptions.NotFoundException;
 import com.example.banco.repositories.ClienteRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -44,6 +45,36 @@ public class ClienteServiceImpl implements ClienteService {
 
     @Override
     public void deleteCliente(Long id) {
-        clienteRepository.deleteById(id);
+        Optional<Cliente> optionalCliente = clienteRepository.findById(id);
+        if (optionalCliente.isPresent()) {
+            clienteRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("El cliente con ID " + id + " no existe.");
+        }
+    }
+    @Override
+    public ClienteDTO updateCliente(Long id, ClienteDTO clienteDTO) {
+        Optional<Cliente> optionalCliente = clienteRepository.findById(id);
+        if (optionalCliente.isPresent()) {
+            Cliente cliente = optionalCliente.get();
+            // Actualiza los campos relevantes con los datos proporcionados en clienteDTO
+            cliente.setNombre(clienteDTO.getNombre());
+            cliente.setApellido(clienteDTO.getApellido());
+            cliente.setDomicilio(clienteDTO.getDomicilio());
+            cliente.setTelefono(clienteDTO.getTelefono());
+            cliente.setDni(clienteDTO.getDni());
+            cliente.setEmail(clienteDTO.getEmail());
+            cliente.setTipoCuenta(clienteDTO.getTipoCuenta());
+
+            // Guarda la entidad actualizada en la base de datos
+            Cliente updatedCliente = clienteRepository.save(cliente);
+
+            // Devuelve el ClienteDTO actualizado
+            return modelMapper.map(updatedCliente, ClienteDTO.class);
+        }
+        return null;
     }
 }
+
+
+
